refactor: extract emails router (PR-4, Phase 1)
Pure move of 4 email endpoints from server/routes.ts into
server/routes/emails.ts, plus the module-level email zod schemas
(emailQuerySchema, markEmailReadSchema) and EMAIL_CATEGORY_VALUES /
emailCategoryEnum, which are used only by these handlers. Mounted via
app.use(emailsRouter) at the original location of the first email
endpoint to preserve registration order. No behavior changes.
